requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
python-dotenv==0.20.0
dj_database_url==1.0.0